requirements.txt
Flask==2.0.1
flask-healthz==0.0.3
gunicorn==20.1.0
honeycomb-beeline==2.17.1
